Moved the projectComponent.js to ProjectsPage.js and adjusted to page layout
