--- denver_agile_metric_calc/src/components/projectComponent.js
@@ -1,26 +1,0 @@
-// projectComponent.js
-import React, { Fragment } from 'react';
-import './ProjectComponent.css';
-import { projects } from './data.js';
-import { getImageUrl } from './utils.js';
-
-export default function ProjectComponent() {
-  const listItems = projects.map(project => (
-    <Fragment key={project.id}>
-      <div className="project-item">
-        <img className="project-image" src={getImageUrl(project)} alt={project.name} />
-        <div className="project-details">
-          <b>{project.name}</b>
-          <p>{project.description}</p>
-        </div>
-      </div>
-    </Fragment>
-  ));
-
-  return (
-    <div>
-      <div className="project-list-heading">User Projects</div>
-      <div className="project-list">{listItems}</div>
-    </div>
-  );
-}
