+++ denver_agile_metric_calc/src/routes/ProjectsPage.js
@@ -1,13 +1,26 @@
-import React from 'react'
+// ProjectsPage.js
+import React, { Fragment } from 'react';
+import '../components/ProjectComponent.css';
+import { projects } from '../components/data.js';
+import { getImageUrl } from '../components/utils.js';
 
-function Projects () {
-    return(
-        <div>
-            <h1>
-                Projects
-            </h1>
+export default function ProjectsPage() {
+  const listItems = projects.map(project => (
+    <Fragment key={project.id}>
+        <div className="project-item">
+            <img className="project-image" src={getImageUrl(project)} alt={project.name} />
+            <div className="project-details">
+            <b>{project.name}</b>
+            <p>{project.description}</p>
+            </div>
         </div>
-    )
-}
+    </Fragment>
+  ));
 
-export default Projects
+  return (
+    <div>
+            <div className="project-list-heading">User Projects</div>
+            <div className="project-list">{listItems}</div>
+    </div>
+  );
+}
